+++ RELEASE_NOTES.md
@@ -3,14 +3,15 @@
 ## Version 1.3.7 - Community 1.0.0
 
     * [SER-3502] - Send country when submitting the account transfer bulk upload
-    
+    * [CP-3538] - Show the transfer description in the view transaction interface if the repayment is from account transfer
+
 ## Version 1.3.6 - Community 1.0.0
 
     * [SER-3363] - Implement label translation in Client related UIs
     * [SER-3259] - External Service: Add new page for payment provider configuration
     * [SER-3364] - Implement label translation in Groups and Loans related UIs
     * [SER-3313] - Add bulk account transaction transfer
-    * [CP-3538] - Show the transfer description in the view transaction interface if the repayment is from account transfer
+
 ## Version 1.3.5 - Community 1.0.0
 
     * [SER-2577] - Allow group movement regardless of loan status and office