+++ RELEASE_NOTES.md
@@ -3,7 +3,8 @@
 ## Version 1.3.7 - Community 1.0.0
 
     * [SER-3502] - Send country when submitting the account transfer bulk upload
-    
+    * [CP-3538] - Show the transfer description in the view transaction interface if the repayment is from account transfer
+
 ## Version 1.3.6 - Community 1.0.0
 
     * [SER-3363] - Implement label translation in Client related UIs